[api][refact] auth
