+++ src/api/auth/LogoutApi.ts
@@ -1,11 +1,11 @@
-import { BaseAPI } from '../core';
+import { BaseAPI } from '../../core';
 
 export class LogoutApi extends BaseAPI {
   constructor(baseUrl?: string) {
     super('/auth/logout', baseUrl);
   }
 
-  async create(): Promise<void> {
+  async create() {
     const data = await this._http.POST(this._url, {});
     console.log('LogoutApi: ', data);
   }
